backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
httpx==0.28.0
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
pgvector==0.3.6

# LLM
anthropic==0.40.0

# Document processing
python-docx==1.1.2
docxtpl==0.19.0
PyMuPDF==1.25.1
pdfplumber==0.11.4
Pillow==11.0.0
pytesseract==0.3.13
openpyxl==3.1.5

# Storage / file handling
aiofiles==24.1.0
python-magic==0.4.27

# Embedding / vector
openai==1.57.4
numpy==2.2.0
scikit-learn==1.6.0
fastembed==0.4.2

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pyyaml==6.0.2
structlog==24.4.0
tenacity==9.0.0
ujson==5.10.0
zipfile36==0.1.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.0
